Directory Categorizer/V8/requirements.txt
google-generativeai>=0.3.0
openai>=1.0.0
cohere>=4.0.0
anthropic>=0.5.0
googlesearch-python>=1.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
